Handle UTF-8 characters.
This doesn't hurt performance at all, and is more correct for the
non-version formatting. It's still not "correct" because we don't handle
non-printable ASCII code (which is fine) and non-alphanumeric sequences
in the 7bit space.

There's still some more work to be done on this front, but hey, it's
closer.